Allow for keyword identifer names
Useful for interop and language extensibility. Apparently, the current ceylon
grammar doesn't allow for keywords in imports even though the specification would,
depending on interpretation.

To resolve keyword/identifier conflicts and interop conflicts, conceptually
distinguish a written identifier from its name. Add two different qualifiers
for LIdentifier and UIdentifier, respectively. Carefully drafted not to allow
the use of $ in an identifier.

V4: use $ and $$ as L/Uidentifier prefixes.

This might make the PIdentifier superfluous.
Disclaimer: not compiled or tested.

Signed-off-by: Simon Thum <simon.thum@gmx.de>