Fix for #518: produce better type casts in case of refined variance inheritance
To fix the fact that in Ceylon subclasses can refine the type parameters of
variant interfaces in subtypes, we used to produce a raw type cast, but that
is not good enough since it loses type annotations. So now we first cast to
a raw type, then to the proper type.