produce the self-type info in @CaseTypes(of) (#283)
Perhaps it'd be a better place to put it in @Class(of) but the spec lists it as a
special case of case types. An alternative would be to put it in @CaseTypes(value)
as a single element and figure it out in the model loader, but I thought this was
clearer.