CodegenUtil: made declName for non-qualified types really non-qualified (#323)
Also updated the tests. If it turns out this fix was wrong, I'll revert and add
a new type of flag but I cna't see why we'd ever need a non-qualified name except
for generating the class name anyways. But the C.C half-qualified notation doesn't
seem to be useful