BoxingDeclarationVisitor: HUGE hack for #689
In order to avoid a ProducedType to be physically shared between a Value and
its Initialiser (in the case of a `value foo = expr;`) we make a copy of the
ProducedType. This allows us to set the underlyingType and isRaw flags of the
expression independently from the Value. IMO this should be done in the typechecker
but this is pretty specific to the compiler…