model loader: made aliases lazy
Turns out I failed to make them non-lazy after all: a toplevel class would load
its inner classes, one of which was an alias, which was eagerly loaded and extended
a sibling inner class, which it tried to find in its container, who wasn't entirely
loaded yet (the other inner classes were yet to be loaded), so it failed. Making
aliases truely lazy solved the issue.