model loader #688: model loader and Throwable
As a result, j.l.Exception and j.l.Error are now subtypes of c.l.Exception
This is a lot saner for everyone concerned. Note that we still don't allow
you to catch more than j.l.Exception, but I'm very tempted to change that
because after all, why prevent people from doing that? That's forcing them
to write Java code if they really want to.