Char/Float/Int/String: added constructors #5114
Note that the parameter of String(String val) is not `string` because that did not compile.
It's not really optimal. I also had to change `formal` methods/attrs in those native declarations
to be `native` so the typechecker would shut up about them not being in an abstract class.
Note sure if that was right. but it works.