remove "compiler bug: nothing captured"
When comprehension clauses can start with an ‘if’ clause, it is possible
to form comprehensions that do not introduce any variables, like this:

    { if (myBoolean) myElement }

In this case, there are no variables, so none can be captured either,
which is not a compiler bug but expected behavior.

See #986 for more information on capturing variables in comprehensions.