Fix issues related to spreading
Fixes #1589 and a follow-up issue reported in #1591.

The compiler would expect the wrong type when transforming a spread
argument – it would always expect the parameter type (Sequential or
Sequence) when only an Iterable type is really required, because
transformArguments will use getSequence() when it encounters a spread,
but not Sequential argument.