Add test for both #5248 and #1640
They are done as part of the LanguageSuite so we don't have to compile
the test suite more than once, even though we end up running it more than
once. These test the classpath tool and use it to run the tests in a new JVM
without JBoss modules or custom class loaders.