TypeParser: Support defaulted types
The TypeLexer can now lex `=` tokens as well (type `EQ`), and the
TypeParser can parse type lists with defaulted types, which can appear
in callable types (`Integer(String,Integer=)`) and tuple types
(`[String,String=]`).

To support this, TypeList.asTuple was completely rewritten; it is now
recursive instead of iterative.

To create union and intersection types, we now use the functions
`unionType` and `intersectionType` instead of `union` and `intersection`
respectively. The difference between them is that the -Type versions
correctly simplify the type (remove duplicates, union with Nothing,
etc.). This was necessary for constructing the tuples in asTuple() since
they would otherwise contain an unnecessary union with Nothing in every
first type parameter to Tuple (Nothing being the element type of
Empty), and also makes sense elsewhere. This shouldn't change anything
in real compiled code, since the types there are already simplified, but
it does break some of the tests in ceylon-compiler, which have to be
updated to take this into account.

This fixes #4579.