Refactor metamodel to support callable and value constructors.
Add CallableConstructorDeclaration (subinterface of FunctionDeclaration) and ValueConstructorDeclaration (subinterface of ValueDeclaration). ConstructorDeclaration becomes an alias for the union of these two.

Add CallableConstructor (subinterface of Function) and ValueConstructor (subinterface of Value).

ClassModel ceases to be Functional (since a Class doesn't already have a parameter list, so cannot have a parameterTypes attribute), but is still Callable (Arguments=Nothing in the case of there being no default constructor).