Fix for #2269: local types not using their qualifying type's type parameters
I had to teach Type about those local types from runtime. In theory I did not change the
code unless we have instances of FunctionOrValueInterface in the qualifying types, which
can only happen at runtime. The typechecker tests all pass, as do the compiler tests,
so this should be good.