Float.integer: show float value in OverflowException
When converting a Float value outside of the Integer range (i.e. to large in either positive or negative direction, an infinity, or NaN) to Integer, you currently get an OverflowException (with the default message "Numeric Overflow"), without any indication which number caused the Overflow.

(In the case which trapped me, I had a NaN, but took quite some debugging time to figure this out.)

This commits adds a slightly more user friendly message to the Exception.

Should we add similar messages to other places where OverflowException is thrown?