Improve message for Character.codepoint
Previous message was `value + " is not a 32-bit integer"`, but the actual check was not for 32-bit-ness, but for `0 < value < #4888FFFF`, so the message could also be thrown for values which actually were 32 bit integers, which was confusing.

Now the message is `value + " is not a possible Unicode code point"`.

I also updated the existing tests for this overflow exception, and added new ones for the border values.