Add error message to Integer.float.
Analogously to the message for Float.integer, this is `value + " cannot be coerced into a 64 bit floating point value"`.

Actually, I'm not sure that this overflow check here is sensible. The class documentation says:

> An integer is considered equal to its `float` representation. That is, for every integer `int`, the expression `int.float==int` evaluates to `true`.

The presence of this exception here means that there are integers (and actually quite the majority, at least in the JVM) where this expression throws an OverflowException instead of evaluating to true (even including some which actually are representable exactly as Float).