Fix problem with comparison with union and intersection types
For Comparable it's possible to have a comparison of type Foo&Integer
or Comparable<Foo|Integer>. This means the lhs gets erased and/or
the argument to comper is not seen to have a compatible type by javac.

Fix this by propagating the type which the expression will be coerced to
to the method which decides the operator optimization strategy.

Also have to fix WithinOp, along similar lines.

Fixes #2330