Analyse control flow in constructor to determine reachability of ctor stmts
This borrows the typechecker's ControlFlowVisitor, changing it to track the
difference between return via normal `return` and return via `throw`
or `assert(false)`.

The compiler has to use a local jump (implemented as a do/while(false)
with labelled breaks) to implement Ceylon's semantics for return in a
constructor. A throw is unchanged, however, and that means javac's
definite return can detect the non-reachability of the post-constructor
initializer statements. So we use the ThrowVisitor to decide whether the
post-constructor statements are reachable, and only copy them into the
constructor if they are.

Fixes #2384