Type Constructors: Various bits and pieces here
* Supporting type parameters contrained to be type constructors
* Supporting subtypes of types with such type parameters
  (when the type parameter is used in a member)
* Making AbstractTypeConstructor instances implement ReifiedType
* Constructing TypeDescriptors for type constructors

The unresolved problem wrt the last bullet is that the reifiedIs() delegates
to the TC which needs to put types in HashSets, which means we need the
TypeAlias to have a Container.

*But* we've never reified local type aliases before, so I was hoping to avoid
using the model loader to load type aliases. Since type constructors are
relatively simple things it seemed like I don't need a "full" model, but it
looks like that's not correct, which means we now need to full model loader
support for local type aliases, so we can obtain a containing declaration.