Fix the ModuleValidator to respect the --target
The ModuleValidator gets the --target from the Context.
When a --target is set we then need to disable the dist overrides mechanism
being apply by CMR. But there's no nice API to do this. For now we kludge it
and disable *all* overrides, using a flag on the ArtifactContext.
To just disable the dist overrides we need to refactor Overrides so it doesn't
just merge all the overrides together.