Fix #5953 reverting the fix for #2423 to fix it another way
In #2423 we see the tc permit assignment of ns (which it can prove must be
null) to String[]?, but javac can't prove the nullness and rejects the
assignment.

This was previously fixed by adding EXPR_DOWNCAST which would explicitly cast
the null expression to the `if` expressions type, but in #5953 we see this is
doesn't work because the `if` expression's type can be wider than the type
being assigned to in the else clause.

So passing either the `else` clause type or the expression type to
transformExpression()/applyEraureAndBoxing() is wrong, depending on the
circumstances. Moreover we've decided to use the else clause type for the
else clause result variable (rather than the overall if expression type)
because of other reasons (#2255).

So in this particular case with null let's just return null explicitly,
having first evaluated the expression which we know have null result
(to obtain any side-effects).