avoid exception creating intersection w/type alias
addToIntersection() may result in a call to TypeDeclaration.isDisjoint()
which does not work correctly for TypeAliases, since isDisjoint() may
call getExtendedType(), which for TypeAliases, returns the type alias's
definition and not the expected (potentially enumerated) extended
supertype.

The patch causes resolveAliases() to be performed prior to
getDeclaration(), thereby eliminating the possibility of the declaration
on which we're calling isDisjoint() being a TypeAlias.

fixes #5981