don't go running away at the first sight of a circular dependency
Processing of imports in verifyModuleDependencyTree() for a module was
aborted upon finding a previously visited import (circular dependency).

But, it's entirely possible for additional imports to exist that have no
yet been visited. This patch replaces 'return' with 'continue' to ensure
all imports are processed.

For example, with the following three modules:

    - A imports B
    - B imports A
    - B imports X

processing of module B's imports { A, X } would return at module A,
leaving module X unprocessed.