Fix #6070
This issue was basically the same thing as #1227 but the nested `for`
loops exposed the inadequacy of the previous fix:
The `DefiniteAssignmentVisitor` was only adding specified `Value`s to the
immediately enclosing `ControlBlock`.

So propagate the specified values up to the outer `for`'s
`ControlBlock` (fixing #6070) That breaks #1227 again because a `return`
effectively exits *all* the enclosing blocks, so for return and throw
we need to iterate up the ControlBlocks and close all pending substitutions.