Switch order of tput invocations
POSIX does not specify how tput works. It usually functions via escape
sequences, which we rely on here. But it's also possible that it works
via syscalls, i.e., always at call time, and can't be caught and
repeated via its output. In that case, we should at least issue the call
that resets the terminal last, so we don't leave the terminal behind in
some weird state.