User $power$ at the value-method corresponding to ^ operator
The problem we're solving here is that, for consistency with other arithmetic
operators, ^ ought to honour small operands. Unlike other arithmetic ops
it is implemented using a method not a JVM opcode. So we need a
`power(int,int)` method. But then, again for consistency with other
operators we've taken the stance that methods in general do *not* honour
small (unless they have small parameters). So when invoking `power` as a
method on small receiver and with small argument you'd get a 32-bit
computation.

The solution used here is to change the value-type method we use for
translating the ^ operator to be $power$, thus there's no power(int,int)
method for javac to select when translating a power method invocation.

I thought this was cleaner than trying to add explicit (long) casts to the
method invocations.

Part of #6038