Fix typed reference used to generate mixed-in DPM
We were passing a typed reference for the parameter we were generating a
DPM for, but that's useless for figuring out the typed parameter for
other (prior) parameters. So pass the parameter explicitly, or pass null when
we're not generating a DPM. This means we don't need DelegatType any more.

Fixes #6213 again