Fix #6285
String.indexesWhere would, on the last character (codepoint), increase
the offset, then compare it to the length, and exit without considering
the last character, incorrectly assuming to have exceeded the string
length. Instead, the offset should be checked before increasing it.
(This also allows us to remove an extra check at the beginning.)