Hack around bogus abstract method added by addAmbiguousMembers()
This was preventing me from compiling the language module using default
methods. What's happening here is that Sequence inherits {Element*}
and {Element+} so by the logic of addAmbiguousMembers()
(because they're not isExactly()) needs abstract methods generated for each
member of Iterable, but Iterable.Absent can only ever be erased to j.l.Object,
so there's not need to add abstract methods in that case.

Part of JDK8 interfaces