Update for #6320 / #6372
Add the dynamic flag to values, and don’t emit a getter for them. (Also
not a setter. So currently setting the value from Ceylon will only set
the value in ex$, not in exports, so the change won’t be visible to
TypeScript. Should be fixed once the whole copying to ex$ becomes
unnecessary.)