optimize Iterators of LazyIterables for #5098
Iterators are now "flattened" when rest is a LazyIterator that delegates
to its own rest (the "root" iterator now calls next() on descendent
iterators directly, bypassing the chain of delegators).

As a result, Iterator.next() becomes an O(1) operation for strings of
LazyIterables, rather than the previous O(n) which would also result in
stack overflows for large n.

This improves iterator performace for Iterables created with
{ x1, x2, ..., xn, *other } (where other is a LazyIterable) and the
default implementation of Iterable.follow().