fix #6397, issue with isTypeArgumentListAssignable
If a covariant type argument is a supertype of the type parameter's
upper bound constraint, we can safely normalize the argument to
Anything. Likewise, we could also choose to normalize it to the upper
bound, or anywhere in between.

By normalizing the type argument in isTypeArgumentListAssignable(), we
can allow, for example, an Inv<out Anything> to be assigned to an
Inv<out String>, if the type parameter's upper bound is String. (We know
an Inv<out Anything> can't actually exist; it will be *at least* an
Inv<String>.)

However, this normalization must be made based on the type parameter's
*upper bound constraint*, and not the type parameter itself, in order
for the test to be effective from within the scope of the type
parameter, if the "argument" we are testing is itself the type
parameter!