Add support for void and any type
For the void type, we ignore the difference to the Anything type. The JS
backend does maintain a void flag, but since there is no difference
between a function declared void and one returning Anything, we don’t
bother with that flag.

For the any type… I don’t think we can do any better than just map it to
Anything. The only way to mimic the TypeScript behavior of it being
assignable to every type would be to map it to Nothing in covariant
locations. And that would make #4285 unsound.