remove memory leak in TypecheckerUnit
A reference to the ModuleSourceMapper was recently added to
TypecheckerUnit, but ModuleSourceMappers hold references to all Modules
used while compiling the unit, including Modules that are not
dependencies of the unit.

Having this reference in TypecheckerUnit causes a memory leak for
applications that retain Modules (which have references to
TypecheckerUnits).