Use getDiagnosticsProducingTypeChecker()
Of course the most obvious name isn’t the right choice, because
getTypeChecker() returns a typechecker that doesn’t produce diagnostics,
and certain functions in the typechecker try to produce diagnostics and
then unconditionally fail (throw) for such a typechecker. I have no idea
how a non-diagnostics-producing typechecker is supposed to be used, but
whatever.