Fix #6952: ceylondoc syntax highlighter breaks on '"'
This moves the char rule before the string rule, and adds an assertion
to the string rule that the character before the opening double quote is
not a single quote. This does not fix all problematic situations – one
example: /* " */ ... /* " */ – but it’s enough to fix this bug, and I
don’t think it should cause regressions in reasonable code.