Wrap all code in HTML in <code>, not just <pre>
This was recommended in Rainbow v1.*, but with v2.* it seems to have
become mandatory: Rainbow appends preloader <div>s to the <pre>, and if
there’s no inner <code> that distinguishes the actual code from those
<div>s, then they get treated like regular code and syntax highlighted.

Unfortunately, this causes the code in the spec to have a very small
font size. I suspect this is a Firefox styling bug, since I wasn’t able
to identify any CSS rule causing this change in the Developer Tools, and
because Chromium doesn’t have this bug. Explicitly setting a fixed font
size seems to fix the bug. (The value of 15.2px is Chromium’s computed
style for the elements.)