assignability condition – fix unfinished sentences
It looks like commit 5887209 (for #3626) broke the grammar here when reorganizing the sections. Previously we had a list with 4 items (where the "except" and "or" made sense), then only two of them remained (the others are now in a later section) and the names changed

This commit moves the "or" between them, and gets rid of the "except".